Parse sdp session version as uint64 (RAV-74)

See merge request soundondigital/ravennakit!3

+++ include/ravennakit/sdp/detail/sdp_origin.hpp
@@ -47,7 +47,7 @@ struct OriginField {
     std::string session_id;
 
     /// The version number for this session description.
-    int session_version {};
+    uint64_t session_version {};
 
     /// Specifies the type of network.
     NetwType network_type {NetwType::undefined};
